experiments/02_resource-constraint/requirements.txt
torch
transformers
accelerate
Pillow
torchvision
click
psutil
